+++ core/database/src/main/kotlin/com/pliniodev/xpimprover/core/database/AppDatabase.kt
@@ -2,8 +2,8 @@ package com.pliniodev.xpimprover.core.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.pliniodev.xpimprover.core.database.data.UserEntity
 import com.pliniodev.xpimprover.core.database.data.UserDao
+import com.pliniodev.xpimprover.core.database.data.UserEntity
 
 @Database(entities = [UserEntity::class], version = 1)
 internal abstract class AppDatabase : RoomDatabase() {
