+++ core/database/src/main/kotlin/com/pliniodev/xpimprover/core/database/data/UserDao.kt
@@ -6,8 +6,8 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 
-@Dao
-internal interface UserDao {
+@Dao 
+interface UserDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertUser(userEntity: UserEntity)
