+++ core/database/src/main/kotlin/com/pliniodev/xpimprover/core/database/di/Di.kt
@@ -3,8 +3,6 @@ package com.pliniodev.xpimprover.core.database.di
 import com.pliniodev.xpimprover.core.database.AppDatabase
 import org.koin.dsl.module
 import androidx.room.Room
-import com.pliniodev.xpimprover.core.database.data.UserCrudUseCaseImpl
-import com.pliniodev.xpimprover.core.database.domain.UserCrudUseCase
 
 private const val DATABASE_NAME = "xp.improver.database"
 
@@ -18,7 +16,4 @@ val databaseModule = module {
     }
     
     single { get<AppDatabase>().userDao() }
-
-    // repository
-    factory<UserCrudUseCase> { UserCrudUseCaseImpl(userDao = get()) }
 }
